fix(auth): use router.navigate instead of window.location.href on 401
Using window.location.href for the post-refresh-failure redirect caused a
full browser reload, discarding the React Query cache, in-flight router
state, and all other in-memory state — a hard reset in a SPA where none
of that is necessary.

Extracted router creation into src/lib/router.ts so api.ts can import
the router instance for programmatic navigation without creating a
circular dependency (main.tsx → api.ts already exists; api.ts → main.tsx
would be a cycle). main.tsx now imports the router from the same module.

Removed the shouldRedirectToLogin() guard alongside the href assignment —
TanStack Router's _auth.tsx layout already prevents the redirect loop for
users already on auth pages.

Documented the router module split and the intentional absence of Zod for
validateSearch in CLAUDE.md to prevent these from being re-raised as
review findings.
